+++ data/migrations/026_supplier_document_sod_guard.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+ALTER TABLE supplier_documents DROP CONSTRAINT IF EXISTS ck_supplier_document_checker;
+COMMIT;